utilize custom_id for infrastructures

+++ lib/models/network.rb
@@ -24,7 +24,7 @@ class Network
   def speed_or_default
     # It's expcted that the user-configured speed be provided in gigabits per second, so we multiply it out to bits per second (for API submission)
     if ENV["DEFAULT_#{kind}_IO"]
-      ENV["DEFAULT_#{kind}_IO"].to_i * 1000000000
+      ENV["DEFAULT_#{kind}_IO"].to_f * 1000000000
     else
       case kind
       when 'WAN' then 1000000000
